scripts/caption/requirements.txt
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
timm>=0.9.0
onnxruntime>=1.15.0
huggingface-hub>=0.16.0
opencv-python>=4.8.0

